+++ src/ops/op_enum.h
@@ -185,6 +185,8 @@ typedef enum {
     E_OP_N_S,
     E_OP_N_C,
     E_OP_N_CS,
+    E_OP_N_B,
+    E_OP_N_BX,
     E_OP_V,
     E_OP_VV,
     E_OP_ER,